pyproject.toml
[project]
name = "scrumagent"
version = "2.0.0"
description = "AI-powered Scrum Master agent with MCP tool integration"
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }
dependencies = [
    # Core Framework
    "langchain>=1.2.8",
    "langchain-core>=1.2.8",
    "langchain-community>=0.4.0",
    "langchain-openai>=1.1.7",
    "langchain-anthropic>=1.3.1",
    "langchain-ollama>=1.0.0",
    "langchain-chroma>=1.0.0",

    # LangGraph
    "langgraph>=1.0.7",
    "langgraph-cli>=0.4.12",
    "langgraph-checkpoint-mongodb>=0.3.0",

    # MCP Integration
    "langchain-mcp-adapters>=0.2.1",
    "langchain-taiga @ git+https://github.com/Shikenso-Analytics/langchain-taiga@main",

    # Discord Bot
    "discord.py>=2.5.0",

    # Web Search Tools
    "ddgs>=9.0.0",
    "arxiv>=2.1.3",
    "wikipedia>=1.4.0",
    "youtube-search>=2.1.2",

    # Storage & Persistence
    "pymongo>=4.11.1",
    "chromadb>=0.6.0",

    # Observability
    "sentry-sdk>=2.0.0",

    # Utilities
    "httpx>=0.28.1",
    "pytz>=2025.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
    "langsmith>=0.3.11",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.1",
    "pytest-asyncio>=0.25.0",
]

[dependency-groups]
dev = [
    "pytest>=8.2.1",
    "pytest-asyncio>=0.25.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
